+++ NEWS.md
@@ -1,3 +1,11 @@
+# PLNmodels 0.9.2.9000
+
+* development version
+
+# PLNmodels 0.9.2
+
+* various bug fixes
+
 # PLNmodels 0.9.1
 
 * Use nloptr to prepare CRAN release